packages/ingest-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
geoip2==4.7.0
maxminddb==2.4.0
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0